-Added configuration to chainquery with file watching and updating -Created default configuration to be committed. -Converted all decimal values in schema to DOUBLE for sqlboiler fix for unit tests. -Added Claims to datastore -Added TableStatus query to get row information on tables in db and unit tests. -Fixed schema bug that made json version 3 look like 2 -Added isUpdate field to claims to support updates. -Added TimeTrack function to call to profile function calls easily. -Updated dev script to remove env variables now that we have configuration file with defaults.

+++ model/input_test.go
@@ -1285,7 +1285,7 @@ func testInputsSelect(t *testing.T) {
 }
 
 var (
-	inputDBTypes = map[string]string{`Coinbase`: `varchar`, `Created`: `datetime`, `ID`: `bigint`, `InputAddressID`: `bigint`, `IsCoinbase`: `tinyint`, `Modified`: `datetime`, `PrevoutHash`: `varchar`, `PrevoutN`: `int`, `PrevoutSpendUpdated`: `tinyint`, `ScriptSigAsm`: `text`, `ScriptSigHex`: `text`, `Sequence`: `int`, `TransactionHash`: `varchar`, `TransactionID`: `bigint`, `Value`: `decimal`}
+	inputDBTypes = map[string]string{`Coinbase`: `varchar`, `Created`: `datetime`, `ID`: `bigint`, `InputAddressID`: `bigint`, `IsCoinbase`: `tinyint`, `Modified`: `datetime`, `PrevoutHash`: `varchar`, `PrevoutN`: `int`, `PrevoutSpendUpdated`: `tinyint`, `ScriptSigAsm`: `text`, `ScriptSigHex`: `text`, `Sequence`: `int`, `TransactionHash`: `varchar`, `TransactionID`: `bigint`, `Value`: `double`}
 	_            = bytes.MinRead
 )
